# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Assignment",
        "version": "0.1"
    },
    "paths": {
        "/company": {
            "post": {
                "operationId": "d7bd964d23e0ec8f55cf3b4e73948627",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "parentCompanyId",
                                    "name"
                                ],
                                "properties": {
                                    "parentCompanyId": {
                                        "type": "integer",
                                        "minimum": 1
                                    },
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/company/{id}": {
            "get": {
                "operationId": "ccc4a01fe867284b8814c5cd10b1f686",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDTO"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Company get",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompanyDTO"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "operationId": "f12718445df7cc2cfe6acc356a0d0da3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id",
                                    "parentCompanyId",
                                    "name"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "minimum": 1
                                    },
                                    "parentCompanyId": {
                                        "type": "integer",
                                        "minimum": 1
                                    },
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDTO"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "operationId": "6e9a19bf383dfcc8ea537a0ed7fe6021",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/company/{id}/closestStations": {
            "get": {
                "operationId": "61e4277768bc812e9ea47d6f61ec3c02",
                "parameters": [
                    {
                        "name": "radius",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "float"
                        }
                    },
                    {
                        "name": "latitude",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "float"
                        }
                    },
                    {
                        "name": "longitude",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "float"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDTO"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Company get closest stations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClosestStationsDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/station": {
            "post": {
                "operationId": "077d5beed0dfc33cf444d097f374b9eb",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "latitude",
                                    "longitude",
                                    "company_id",
                                    "name",
                                    "address"
                                ],
                                "properties": {
                                    "latitude": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "companyId": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/station/{id}": {
            "get": {
                "operationId": "55edcc6f311195205a47b8be5abaa945",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDTO"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Company get",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StationDTO"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "operationId": "291d652d1791077eb00185d10baec379",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id",
                                    "latitude",
                                    "longitude",
                                    "company_id",
                                    "name",
                                    "address"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "companyId": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDTO"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "operationId": "e839ba82d870d31d84a7b25ed73d62e8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDTO"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "CloseStationDTO": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string"
                    },
                    "lat": {
                        "type": "number",
                        "format": "float"
                    },
                    "long": {
                        "type": "number",
                        "format": "float"
                    },
                    "distance": {
                        "type": "number",
                        "format": "float"
                    }
                },
                "type": "object"
            },
            "ClosestStationsDTO": {
                "properties": {
                    "stations": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CloseStationDTO"
                        }
                    }
                },
                "type": "object"
            },
            "CompanyDTO": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "parentCompanyId": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ErrorDTO": {
                "properties": {
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "StationDTO": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "latitude": {
                        "type": "number",
                        "format": "float"
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float"
                    },
                    "address": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    }
}